+++ Scripts/setup_venv.sh
@@ -2,6 +2,11 @@
 
 set -e
 
+# This is to take python2 from pyenv
+export PYENV_ROOT="$HOME/.pyenv"
+export PATH="$PYENV_ROOT/bin":$PATH
+eval "$(pyenv init -)" || echo "No pyenv, procceding without"
+
 dockerImage=$1
 name=$2
 dirPath=$3
@@ -12,7 +17,7 @@ pythonPath=$5
 export PATH=/opt/homebrew/bin:$PATH
 
 cd "$dirPath"
-testImage=$(docker images --format "{{.Repository}}:{{.Tag}}" | grep devtest"$dockerImage" | head -1)
+testImage=$(docker images --format "{{.Repository}}:{{.Tag}}" | grep "$dockerImage" | head -1)
 echo "Using test image env: $testImage"
 docker rm -f "${name}" &> /dev/null || true
 pythonVersion=$(docker run --name ${name} ${testImage} "python -c 'import sys; print(sys.version_info[0])'")
@@ -22,10 +27,19 @@ docker run --name "${name}" "$testImage" 'pip list --format=freeze > /requiremen
 docker cp "${name}":/requirements.txt .
 docker rm -f "${name}" || true
 
+# check if virtualenv module is installed
+isVirtualEnvInstalled=true
+$pythonPath -m virtualenv --version > /dev/null 2>&1 || isVirtualEnvInstalled=false
+if [ "$isVirtualEnvInstalled" = "false" ]; then
+    $pythonPath -m pip install virtualenv
+fi
 $pythonPath -m virtualenv -p python"${pythonVersion}" venv
-# install all dependency one by one. If one or more fails, we continue.
-while read line; do venv/bin/pip install --disable-pip-version-check "$line" || echo "Could not install dependency $line, proceeding"; done < requirements.txt
-venv/bin/pip install autopep8 --disable-pip-version-check
+venv/bin/pip --version || (echo "No pip, check your python"${pythonVersion}" installation" && exit 1)
+while read line; do
+    venv/bin/pip install --disable-pip-version-check --no-cache-dir $line || echo "Could not install dependency $line, proceeding"
+done < requirements.txt
+venv/bin/pip install autopep8 --disable-pip-version-check --no-cache-dir || echo "Could not install autopep8"
+venv/bin/pip install flake8 --disable-pip-version-check --no-cache-dir || echo "Could not install flake8"
 if [ "${pythonVersion}" = "3" ]; then
-    venv/bin/pip install -r "$extraReqs" --disable-pip-version-check
+    venv/bin/pip install -r "$extraReqs" --disable-pip-version-check --no-cache-dir || echo "Could not install mypy"
 fi